requirements.txt
rich>=13.0.0
requests>=2.25.0
requests-toolbelt>=0.10.0
python-dotenv>=0.19.0 